Title: 삼 오 무, Time: 162ms, Memory: 10MB, Status: Passed - CodeTree

+++ 241015/삼 오 무/three-five-moo.java
@@ -2,14 +2,14 @@ import java.util.*;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int N = sc.nextInt();
+        long N = sc.nextLong();
 
-        int ans = Integer.MAX_VALUE;
-        int left = 1;
-        int right = Integer.MAX_VALUE;
+        long ans = Long.MAX_VALUE;
+        long left = 1;
+        long right = Long.MAX_VALUE;
 
         while(left <= right){
-            int mid = (left + right) / 2;
+            long mid = (left + right) / 2;
             if(mooNum(mid) >= N){
                 right = mid -1;
                 ans = Math.min(ans, mid);
@@ -20,8 +20,8 @@ public class Main {
 
     }
 
-    static int mooNum(int mid){
-        int moo = mid / 3 + mid / 5 - mid / 15;
+    static long mooNum(long mid){
+        long moo = mid / 3 + mid / 5 - mid / 15;
         return mid - moo;
     }
 }
